+++ PandaSQL/Utils/Bitmask.cpp
@@ -7,7 +7,7 @@
 namespace PandaSQL
 {
 
-Bitmask::Bitmask(uint32_t inLength)
+Bitmask::Bitmask(size_t inLength)
 :
 mLength(inLength)
 ,mpBits(new bool[mLength])
@@ -21,25 +21,25 @@ Bitmask::~Bitmask()
 	mpBits = NULL;
 }
 
-bool Bitmask::GetBit(uint32_t index) const
+bool Bitmask::GetBit(size_t index) const
 {
 	PDASSERT(index < mLength);
 
 	return mpBits[index];
 }
 
-void Bitmask::SetBit(uint32_t index, bool bit)
+void Bitmask::SetBit(size_t index, bool bit)
 {
 	PDASSERT(index < mLength);
 	mpBits[index] = bit;
 }
 
-void Bitmask::SetBit(uint32_t low, uint32_t high, bool bit)
+void Bitmask::SetBit(size_t low, size_t high, bool bit)
 {
 	PDASSERT(low <= high);
 	PDASSERT(high < mLength);
 
-	uint32_t index;
+	size_t index;
 	for (index = low; index <= high; index++)
 	{
 		mpBits[index] = bit;
@@ -50,7 +50,7 @@ bool Bitmask::IsClear() const
 {
 	bool result = true;
 
-	uint32_t i;
+	size_t i;
 	for (i = 0; i < this->GetLength(); i++)
 	{
 		if (GetBit(i))
@@ -73,7 +73,7 @@ void Bitmask::Union(const Bitmask &left, const Bitmask &right)
 	PDASSERT(left.GetLength() == right.GetLength());
 	PDASSERT(this->GetLength() == left.GetLength());
 
-	uint32_t i;
+	size_t i;
 	for (i = 0; i < this->GetLength(); i++)
 	{
 		this->SetBit(i, left.GetBit(i) | right.GetBit(i));
