+++ PandaSQL/Utils/Bitmask.h
@@ -11,22 +11,22 @@ class Bitmask
 
 public:
 
-	Bitmask(uint32_t inLength);
+	Bitmask(size_t inLength);
 	~Bitmask();
 
-	bool GetBit(uint32_t index) const;
-	void SetBit(uint32_t index, bool bit);
-	void SetBit(uint32_t low, uint32_t high, bool bit);
+	bool GetBit(size_t index) const;
+	void SetBit(size_t index, bool bit);
+	void SetBit(size_t low, size_t high, bool bit);
 	
 	bool IsClear() const;
 	void ClearAll();
-	uint32_t GetLength() const { return mLength; }
+	size_t GetLength() const { return mLength; }
 
 	void Union(const Bitmask &left, const Bitmask &right);
 	
 private:
 
-	uint32_t mLength;
+	size_t mLength;
 	bool *mpBits;
 
 };
